requirements.txt
discord.py==2.5.0
yt-dlp==2025.2.19
pydub==0.25.1
numpy==1.26.4
aiohttp==3.9.5
requests==2.32.2
python-dotenv==1.0.1
ffmpeg==1.4
PyNaCl==1.5.0
asyncpg==0.29.0
fuzzywuzzy
python-Levenshtein
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1